separate the tests
